// demo-mcp-21-code-generator-node/input/reference-repos/ECOMMERCE_V9/MSAbCaAlarFeaturesETL/src/common/utils/enums/taks.enum.ts
/**
*Enumera las tareas que se realizan en el microservicio
*@author  Santiago Vargas
*/

//Enumera las tareas que se realizan en el microservicio
export enum Etask {
    CREATE = 'PROCESS_CREATE',
    CHANNEL = 'VALIDATE_CHANNEL',
    LOAD_MESSAGE = 'CARGANDO_MENSAJES',
    REQUEST_HTTP = 'REQUEST_HTTS',
    APM='CAPTURE_APM_QUERY',
    ERROR_LOAD_PARAM = 'ERROR_CARGANDO_PARAMETROS',
    CONSUMED_SERVICE = 'CONSUMED_SERVICE',
    VALIDATE_REQUEST='VALIDATE_REQUEST',
    MAP_FEATURES = "MAPEANDO CARACTERISTICAS",
    MAP_NAME = "MAPEANDO NAME",
    MAP_URLKEYWORD = "MAPEANDO URLKEYWORD",
    UPDATE_FEATURES = "ACTUALIZANDO CARACTERISTICAS",
    START_UPDATE_FEATURES = "INICIA ACTUALIZANDO CARACTERISTICAS",
    FINISH_UPDATE_FEATURES = "FINALIZA ACTUALIZANDO CARACTERISTICAS",
    INSERT_FEATURES = "INSERTANDO CARACTERISTICAS",
    DELETE_FEATURES = "BORRANDO CARACTERISTICAS",
    CREATE_IMAGES = "CREANDO IMAGENES",
    EXEC_CRON_CATEGORIES = "EJECUCIÓN CRON CATEGORIAS",
    END_UPDATE_FEATURES = "FIN ACTUALIZACIÓN CARACTERISTICAS",
    EMIT_KAFKA = 'EMITIENDO EVENTO KAFKA',
    FIND_CONTINGENCY = 'CONSULTANDO DATA CONTINGENCIA',
    FIND_AVAILABLE_FEATURES = 'CONSULTANDO CARACTERISTICAS HABILITADAS',
    FIND_PRICES = 'CONSULTANDO PRECIOS',
    CROSS_DATA = 'CRUZANDO INFORMACIÓN',
    LIST_PARENTS = "CONFIGURACION PADRES-HIJOS",
    SAVE_LIST_PARENTS = 'GUARDANDO PADRES-HIJOS',
    FIND_DATA = 'CONSULTANDO CATALOGOS',
    CREATE_NAME_FOR_TECHNOLOGY = 'CREATE NAME FOR TECHNOLOGY',
    ERROR_TRACE = 'ERROR IN TRACEABILITY',
    UPDATE_REGULATORY_FEATURES = 'UPDATE REGULATORY FEATURES'
}

//Descripción de las tareas que se realizan en el microservicio
export enum ETaskDesc {
    CHANNEL = 'Validation of the channel',
    ERROR_LOAD_PARAM = 'Error cargando parametros',
    ERROR_MAP_FEATURES = "ERROR MAPENADO CARACTERISTICAS",
    ERROR_MAP_NAME_FEATURES = "ERROR ACTUALIZANDO NAME",
    ERROR_MAP_URLKEYWORD_FEATURES = "ERROR ACTUALIZANDO URLKEYWORD",
    ERROR_UPDATE_FEATURES = "ERROR ACTUALIZANDO CARACTERISTICAS",
    ERROR_CREATE_NAME_FOR_TECHNOLOGY = 'Error creando el nombre para tecnologia',
    ERROR_INSERT_FEATURES = "ERROR INSERTANDO CARACTERISTICAS",
    ERROR_DELETE_FEATURES = "ERROR BORRANDO CARACTERISTICAS",
    START_UPDATE_FEATURES = "Inicia el proceso de actualización de caracteristicas",
    FINAL_FEATURES_UPDATE = "Caracteristicas actualizadas",
    START_FEATURES_UPDATE_JOB = "Inicia el mapeo de caracteristicas desde el job",
    ERROR_EMIT_KAFKA = 'Error emitiendo evento kafka',
    ERROR_CROSS_DATA = 'Error cruzando información de precios',
    ERROR_GROUP_PRODUCTS = 'Error agrupando productos',
    ERROR_SAVE_LIST_PARENTS = 'Error guardando lista de padres',
    ERROR_FIND_DATA = 'Error consultando información',
}

export enum ETaskTrace {
    FAILED ="FAILED",
    SUCCESS="SUCCESS"
}